MAMS-backend/requirements.txt
Flask==2.2.5
flask_bcrypt==1.0.1
flask_cors==5.0.1
flask_jwt_extended==4.7.1
flask_sqlalchemy==3.1.1
psycopg2==2.9.10
PyJWT==2.8.0
PyJWT==2.10.1
python-dotenv==1.0.1
supabase==2.13.0
Werkzeug==3.1.3
